+++ Tests/JsonArrayTestsGroup.cpp
@@ -175,7 +175,7 @@ TEST(JsonArrayTestsGroup, JsonObjectArray_WriteTo_With_Limited_Buffer_Test) {
 
 static void *TestParent = NULL;
 static char *DirectWriteTestBuffer = NULL;
-static void OnReady(void *parent, const char *json, int size) {
+static void OnReady(void *parent, const char *json, size_t size) {
 	TestParent = parent;
 	DirectWriteTestBuffer = new char[size + 1];
 	memcpy(DirectWriteTestBuffer, json, size);
